pyproject.toml
[project]
name = "tg-bot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.27.0",
    "python-dotenv>=1.2.2",
]
